requirements.txt
python-vlc
paho-mqtt
RPi.GPIO
toml
pyALSAAUDIO
